+++ README.md
@@ -107,17 +107,24 @@ website is using Next.js SSG and Tailwind CSS for styling.
 
 ### Testing grounds
 
+#### DOM
+
 `npm -w packages/dom run dev` in the root will launch the `@floating-ui/dom`
 development visual tests at `http://localhost:1234`. The playground uses React
-to write each test route, bundled by Parcel.
+to write each test route, bundled by Vite.
 
 Each route has screenshots taken of the page by Playwright to ensure all the
 functionalities work as expected; this is an easy, reliable and high-level way
-of testing the code.
+of testing the positioning logic.
 
 Below the main container are UI controls to turn on certain state and options.
 Every single combination of state is tested visually via the snapshots to cover
 as much as possible.
+
+#### React
+
+`npm -w packages/react run dev` in the root will launch the `@floating-ui/react`
+development tests at `http://localhost:1234`.
 
 ## README Sponsors
