--- packages/core/src/_babel.d.ts
@@ -1,1 +1,0 @@
-declare const __DEV__: boolean;
