+++ packages/dom/src/platform.ts
@@ -1,34 +1,24 @@
+import {convertOffsetParentRelativeRectToViewportRelativeRect} from './platform/convertOffsetParentRelativeRectToViewportRelativeRect';
+import {getClientRects} from './platform/getClientRects';
+import {getClippingRect} from './platform/getClippingRect';
+import {getDimensions} from './platform/getDimensions';
+import {getDocumentElement} from './platform/getDocumentElement';
+import {getElementRects} from './platform/getElementRects';
+import {getOffsetParent} from './platform/getOffsetParent';
+import {getScale} from './platform/getScale';
+import {isElement} from './platform/isElement';
+import {isRTL} from './platform/isRTL';
 import type {Platform} from './types';
-import {convertOffsetParentRelativeRectToViewportRelativeRect} from './utils/convertOffsetParentRelativeRectToViewportRelativeRect';
-import {getClippingRect} from './utils/getClippingRect';
-import {getComputedStyle} from './utils/getComputedStyle';
-import {getDimensions} from './utils/getDimensions';
-import {getDocumentElement} from './utils/getDocumentElement';
-import {getOffsetParent} from './utils/getOffsetParent';
-import {getRectRelativeToOffsetParent} from './utils/getRectRelativeToOffsetParent';
-import {getScale} from './utils/getScale';
-import {isElement} from './utils/is';
 
 export const platform: Required<Platform> = {
-  getClippingRect,
   convertOffsetParentRelativeRectToViewportRelativeRect,
-  isElement,
-  getDimensions,
-  getOffsetParent,
   getDocumentElement,
+  getClippingRect,
+  getOffsetParent,
+  getElementRects,
+  getClientRects,
+  getDimensions,
   getScale,
-  async getElementRects({reference, floating, strategy}) {
-    const getOffsetParentFn = this.getOffsetParent || getOffsetParent;
-    const getDimensionsFn = this.getDimensions;
-    return {
-      reference: getRectRelativeToOffsetParent(
-        reference,
-        await getOffsetParentFn(floating),
-        strategy
-      ),
-      floating: {x: 0, y: 0, ...(await getDimensionsFn(floating))},
-    };
-  },
-  getClientRects: (element) => Array.from(element.getClientRects()),
-  isRTL: (element) => getComputedStyle(element).direction === 'rtl',
+  isElement,
+  isRTL,
 };
