+++ packages/dom/src/platform/getClientRects.ts
@@ -1,0 +1,3 @@
+export function getClientRects(element: Element) {
+  return Array.from(element.getClientRects());
+}
