+++ packages/dom/src/platform/getDimensions.ts
@@ -1,6 +1,6 @@
 import type {Dimensions} from '@floating-ui/core';
 
-import {getCssDimensions} from './getCssDimensions';
+import {getCssDimensions} from '../utils/getCssDimensions';
 
 export function getDimensions(element: Element): Dimensions {
   return getCssDimensions(element);
