+++ packages/dom/src/platform/isElement.ts
@@ -1,0 +1,5 @@
+import {getWindow} from '../utils/getWindow';
+
+export function isElement(value: any): value is Element {
+  return value instanceof getWindow(value).Element;
+}
