+++ packages/dom/src/platform/isRTL.ts
@@ -1,0 +1,3 @@
+export function isRTL(element: Element) {
+  return getComputedStyle(element).direction === 'rtl';
+}
