+++ packages/dom/src/utils/getRectRelativeToOffsetParent.ts
@@ -1,11 +1,11 @@
 import type {Rect, Strategy, VirtualElement} from '@floating-ui/core';
 
+import {getDocumentElement} from '../platform/getDocumentElement';
 import {getBoundingClientRect} from './getBoundingClientRect';
-import {getDocumentElement} from './getDocumentElement';
 import {getNodeScroll} from './getNodeScroll';
 import {getWindowScrollBarX} from './getWindowScrollBarX';
 import {isHTMLElement, isOverflowElement} from './is';
-import {createEmptyCoords} from './math';
+import {createCoords} from './math';
 import {getNodeName} from './node';
 
 export function getRectRelativeToOffsetParent(
@@ -19,7 +19,7 @@ export function getRectRelativeToOffsetParent(
   const rect = getBoundingClientRect(element, true, isFixed, offsetParent);
 
   let scroll = {scrollLeft: 0, scrollTop: 0};
-  const offsets = createEmptyCoords(0);
+  const offsets = createCoords(0);
 
   if (isOffsetParentAnElement || (!isOffsetParentAnElement && !isFixed)) {
     if (
