+++ packages/dom/src/utils/math.ts
@@ -2,4 +2,4 @@ export const min = Math.min;
 export const max = Math.max;
 export const round = Math.round;
 export const floor = Math.floor;
-export const createEmptyCoords = (v: number) => ({x: v, y: v});
+export const createCoords = (v: number) => ({x: v, y: v});
