+++ packages/dom/src/utils/unwrapElement.ts
@@ -1,5 +1,5 @@
+import {isElement} from '../platform/isElement';
 import type {VirtualElement} from '../types';
-import {isElement} from './is';
 
 export function unwrapElement(element: Element | VirtualElement) {
   return !isElement(element) ? element.contextElement : element;
